Renaming a file
In your repository, browse to the file you want to rename.
Edit file icon In the upper right corner of the file view, click  to open the file editor.
Editing a file name In the filename field, change the name of the file to the new filename you want. You can also update the contents of your file at the same time.